refactor: reorganize cluster test suite by moving distance matrix tests to a new file and grouping existing tests into classes.
