+++ tests/cluster/test_cluster.py
@@ -4,132 +4,83 @@ import numpy as np
 from falcon.cluster import cluster, distance_matrix
 
 
-def test_get_precursor_mz_splits_single_split():
-    """Test the case where there is only one split."""
-    precursor_mzs = np.array([100.0, 100.5, 101.0, 101.5])
-    splits = cluster._get_precursor_mz_splits(
-        precursor_mzs, 0.5, "Da", batch_size=100
-    )
-    assert splits == nb.typed.List([0, 4])
+class TestGetPrecursorMzSplits:
+    def test_get_precursor_mz_splits_single_split(self):
+        """Test the case where there is only one split."""
+        precursor_mzs = np.array([100.0, 100.5, 101.0, 101.5])
+        splits = cluster._get_precursor_mz_splits(
+            precursor_mzs, 0.5, "Da", batch_size=100
+        )
+        assert splits == nb.typed.List([0, 4])
+
+    def test_get_precursor_mz_splits_multiple_splits(self):
+        """Test the case where there are multiple splits."""
+        precursor_mzs = np.array([100.0, 100.5, 101.5, 102.0])
+        splits = cluster._get_precursor_mz_splits(
+            precursor_mzs, 0.5, "Da", batch_size=100
+        )
+        assert splits == nb.typed.List([0, 2, 4])
+
+    def test_get_precursor_mz_splits_single_split_ppm(self):
+        """Test the case where there is only one split with ppm tolerance."""
+        precursor_mzs = np.array([100.0, 100.002, 100.003, 100.005])
+        splits = cluster._get_precursor_mz_splits(
+            precursor_mzs, 50, "ppm", batch_size=100
+        )
+        assert splits == nb.typed.List([0, 4])
+
+    def test_get_precursor_mz_splits_multiple_splits_ppm(self):
+        """Test the case where there are multiple splits with ppm tolerance."""
+        precursor_mzs = np.array([100.0, 100.005, 101.5, 101.005])
+        splits = cluster._get_precursor_mz_splits(
+            precursor_mzs, 50, "ppm", batch_size=100
+        )
+        assert splits == nb.typed.List([0, 2, 4])
+
+    def test_get_precursor_mz_splits_batch_size(self):
+        """Test the case where the batch size is smaller than the number of precursor m/z values."""
+        precursor_mzs = np.array([100.0, 100.2, 100.4, 100.5])
+        splits = cluster._get_precursor_mz_splits(
+            precursor_mzs, 0.5, "Da", batch_size=2
+        )
+        assert splits == nb.typed.List([0, 2, 4])
+
+        precursor_mzs = np.array([100, 101, 102, 103, 106, 107])
+        splits = cluster._get_precursor_mz_splits(
+            precursor_mzs, 1, "Da", batch_size=3
+        )
+        assert splits == nb.typed.List([0, 3, 4, 6])
+
+    def test_get_precursor_mz_splits_empty(self):
+        """Test the case where the precursor m/z values are empty."""
+        precursor_mzs = np.array([])
+        splits = cluster._get_precursor_mz_splits(
+            precursor_mzs, 0.5, "Da", batch_size=100
+        )
+        assert splits == nb.typed.List([0])
+
+    def test_get_precursor_mz_splits_single_value(self):
+        """Test the case where there is only one precursor m/z value."""
+        precursor_mzs = np.array([100.0])
+        splits = cluster._get_precursor_mz_splits(
+            precursor_mzs, 0.5, "Da", batch_size=100
+        )
+        assert splits == nb.typed.List([0, 1])
 
 
-def test_get_precursor_mz_splits_multiple_splits():
-    """Test the case where there are multiple splits."""
-    precursor_mzs = np.array([100.0, 100.5, 101.5, 102.0])
-    splits = cluster._get_precursor_mz_splits(
-        precursor_mzs, 0.5, "Da", batch_size=100
-    )
-    assert splits == nb.typed.List([0, 2, 4])
+class TestLinkage:
+    def test_linkage_da(self):
+        """Test the linkage function with Da tolerance."""
+        values = np.array([100, 101, 105, 107, 110])
+        linkage_matrix = cluster._linkage(values, "Da")
+        assert linkage_matrix.shape == (4, 4)
+        print(linkage_matrix)
+        assert (
+            linkage_matrix
+            == np.array([[0, 1, 1, 2], [2, 3, 2, 2], [6, 4, 5, 3], [5, 7, 10, 5]])
+        ).all()
 
 
-def test_get_precursor_mz_splits_single_split_ppm():
-    """Test the case where there is only one split with ppm tolerance."""
-    precursor_mzs = np.array([100.0, 100.002, 100.003, 100.005])
-    splits = cluster._get_precursor_mz_splits(
-        precursor_mzs, 50, "ppm", batch_size=100
-    )
-    assert splits == nb.typed.List([0, 4])
-
-
-def test_get_precursor_mz_splits_multiple_splits_ppm():
-    """Test the case where there are multiple splits with ppm tolerance."""
-    precursor_mzs = np.array([100.0, 100.005, 101.5, 101.005])
-    splits = cluster._get_precursor_mz_splits(
-        precursor_mzs, 50, "ppm", batch_size=100
-    )
-    assert splits == nb.typed.List([0, 2, 4])
-
-
-def test_get_precursor_mz_splits_batch_size():
-    """Test the case where the batch size is smaller than the number of precursor m/z values."""
-    precursor_mzs = np.array([100.0, 100.2, 100.4, 100.5])
-    splits = cluster._get_precursor_mz_splits(
-        precursor_mzs, 0.5, "Da", batch_size=2
-    )
-    assert splits == nb.typed.List([0, 2, 4])
-
-    precursor_mzs = np.array([100, 101, 102, 103, 106, 107])
-    splits = cluster._get_precursor_mz_splits(
-        precursor_mzs, 1, "Da", batch_size=3
-    )
-    assert splits == nb.typed.List([0, 3, 4, 6])
-
-
-def test_get_precursor_mz_splits_empty():
-    """Test the case where the precursor m/z values are empty."""
-    precursor_mzs = np.array([])
-    splits = cluster._get_precursor_mz_splits(
-        precursor_mzs, 0.5, "Da", batch_size=100
-    )
-    assert splits == nb.typed.List([0])
-
-
-def test_get_precursor_mz_splits_single_value():
-    """Test the case where there is only one precursor m/z value."""
-    precursor_mzs = np.array([100.0])
-    splits = cluster._get_precursor_mz_splits(
-        precursor_mzs, 0.5, "Da", batch_size=100
-    )
-    assert splits == nb.typed.List([0, 1])
-
-
-def test_linkage_da():
-    """Test the linkage function with Da tolerance."""
-    values = np.array([100, 101, 105, 107, 110])
-    linkage_matrix = cluster._linkage(values, "Da")
-    assert linkage_matrix.shape == (4, 4)
-    print(linkage_matrix)
-    assert (
-        linkage_matrix
-        == np.array([[0, 1, 1, 2], [2, 3, 2, 2], [6, 4, 5, 3], [5, 7, 10, 5]])
-    ).all()
-
-
-def test_condensed_index_basic():
-    """Test basic cases where the condensed index is computed correctly."""
-    assert distance_matrix.condensed_index(0, 1, 5) == 0
-    assert distance_matrix.condensed_index(0, 2, 5) == 1
-    assert distance_matrix.condensed_index(1, 3, 5) == 5
-    assert distance_matrix.condensed_index(2, 3, 5) == 7
-    assert distance_matrix.condensed_index(3, 4, 5) == 9
-
-
-def test_condensed_index_swapped_inputs():
-    """Ensure (i, j) gives the same result as (j, i)."""
-    assert distance_matrix.condensed_index(2, 4, 5) == distance_matrix.condensed_index(4, 2, 5)
-    assert distance_matrix.condensed_index(0, 3, 5) == distance_matrix.condensed_index(3, 0, 5)
-
-
-def test_condensed_index_invalid_diagonal():
-    """Check that passing (i, i) raises a ValueError."""
-    with pytest.raises(
-        ValueError, match="No diagonal elements in condensed matrix"
-    ):
-        distance_matrix.condensed_index(2, 2, 5)
-
-
-def test_condensed_index_out_of_bounds():
-    """Test cases where i or j is out of range (negative or >= n)."""
-    with pytest.raises(ValueError):
-        distance_matrix.condensed_index(-1, 2, 5)
-    with pytest.raises(ValueError):
-        distance_matrix.condensed_index(2, 5, 5)
-    with pytest.raises(ValueError):
-        distance_matrix.condensed_index(5, 2, 5)
-
-
-def test_condensed_index_invalid_n():
-    """Test cases where n is invalid."""
-    with pytest.raises(ValueError):
-        distance_matrix.condensed_index(0, 1, 0)
-    with pytest.raises(ValueError):
-        distance_matrix.condensed_index(0, 1, -1)
-
-
-def test_condensed_index_large_matrix():
-    """Test with a larger matrix to check correct indexing."""
-    assert distance_matrix.condensed_index(10, 20, 100) == 954
-    assert distance_matrix.condensed_index(50, 99, 100) == 3773
 
 
 # ---------------------------------------------------------------------------
@@ -247,33 +198,4 @@ class TestCostBasedChunking:
         assert len(chunks[0]) == 3
 
 
-# ---------------------------------------------------------------------------
-# P3: compute_condensed_distance_matrix
-# ---------------------------------------------------------------------------
 
-
-class TestComputeCondensedDistanceMatrix:
-    def test_identical_spectra(self, spectrum_tuple_pair):
-        """Identical spectra should have distance 0."""
-        spec1, spec2 = spectrum_tuple_pair
-        pdist = cluster.compute_condensed_distance_matrix(
-            [spec1, spec2], fragment_mz_tol=0.5, min_matches=0
-        )
-        assert abs(pdist[0]) < 0.01
-
-    def test_orthogonal_spectra(self, orthogonal_spectrum_tuple_pair):
-        """Non-overlapping spectra should have distance ~1."""
-        spec1, spec2 = orthogonal_spectrum_tuple_pair
-        pdist = cluster.compute_condensed_distance_matrix(
-            [spec1, spec2], fragment_mz_tol=0.05, min_matches=0
-        )
-        assert pdist[0] > 0.99
-
-    def test_min_matches_filter(self, spectrum_tuple_pair):
-        """Pairs with fewer matched peaks than min_matches should get distance 1."""
-        spec1, spec2 = spectrum_tuple_pair
-        # spec1 and spec2 have 3 peaks each; setting min_matches=100 should force dist=1
-        pdist = cluster.compute_condensed_distance_matrix(
-            [spec1, spec2], fragment_mz_tol=0.5, min_matches=100
-        )
-        assert abs(pdist[0] - 1.0) < 0.01
